requirements.txt
agno
openai
groq
python-dotenv
yfinance
lancedb
pandas
duckduckgo-search